auth/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pyjwt==2.8.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
bcrypt==4.1.2
redis==5.0.1